Edited Stuff

+++ partial_diffeq.hpp
@@ -180,10 +180,10 @@ init_vectors(const unsigned short input_points)
     {
       temp_vector[i] += Bottom_Function(x_position);
     }
-
     //(1/M) * (sum of 4 surrounding) - (h^2 / M) * forcing function
     temp_vector[i] /= input_points;
-    temp_vector[i] -= ((pow((1.0 / input_points), 2) / input_points) * y_position);
+    temp_vector[i] -= ((pow((1 / static_cast<float>(input_points)), 2) / 
+                        input_points) * y_position);
   }
 
   m_b_vector = temp_vector;
